added environment variables for clear video and record video

+++ conditions/clear_videos.py
@@ -1,10 +1,12 @@
 import requests
 from lxml import html
+import os
 
-ip = "http://10.10.20.39"
-page = requests.get(f'{ip}:8080/video')
-webpage = html.fromstring(page.content)
-names = webpage.xpath('//a/@href')
-for i in names:
-    url = f'{ip}:4444/video/{i}'
-    requests.delete(url)
+if os.getenv('CLEAR_VIDEOS') == 'true':
+    ip = os.getenv('SELENOID_IP', '127.0.0.1')
+    page = requests.get(f'http://{ip}:8080/video')
+    webpage = html.fromstring(page.content)
+    names = webpage.xpath('//a/@href')
+    for i in names:
+        url = f'http://{ip}:4444/video/{i}'
+        requests.delete(url)
